// packages/contractkit/src/generated/DoubleSigningSlasher.ts
interface EventOptions {
  filter?: object
  fromBlock?: BlockType
  topics?: string[]
}

interface ContractEventLog<T> extends EventLog {
  returnValues: T
}

interface ContractEventEmitter<T> extends EventEmitter {
  on(event: 'connected', listener: (subscriptionId: string) => void): this
  on(event: 'data' | 'changed', listener: (event: ContractEventLog<T>) => void): this
  on(event: 'error', listener: (error: Error) => void): this
}

type ContractEvent<T> = (
  options?: EventOptions,
  cb?: Callback<ContractEventLog<T>>
) => ContractEventEmitter<T>

export interface DoubleSigningSlasher extends Contract {
  clone(): DoubleSigningSlasher
  methods: {
    slashingIncentives(): TransactionObject<{
      penalty: string
      reward: string
      0: string
      1: string
    }>

    validatorSignerAddressFromCurrentSet(index: number | string): TransactionObject<string>

    initialized(): TransactionObject<boolean>

    checkProofOfPossession(
      sender: string,
      blsKey: string | number[],
      blsPop: string | number[]
    ): TransactionObject<boolean>

    getEpochNumberOfBlock(blockNumber: number | string): TransactionObject<string>

    getVerifiedSealBitmapFromHeader(header: string | number[]): TransactionObject<string>

    validatorSignerAddressFromSet(
      index: number | string,
      blockNumber: number | string
    ): TransactionObject<string>

    hashHeader(header: string | number[]): TransactionObject<string>

    renounceOwnership(): TransactionObject<void>

    minQuorumSizeInCurrentSet(): TransactionObject<string>

    registry(): TransactionObject<string>

    numberValidatorsInCurrentSet(): TransactionObject<string>

    groupMembershipAtBlock(
      validator: string,
      blockNumber: number | string,
      groupMembershipHistoryIndex: number | string
    ): TransactionObject<string>

    getBlockNumberFromHeader(header: string | number[]): TransactionObject<string>

    owner(): TransactionObject<string>

    isOwner(): TransactionObject<boolean>

    getEpochNumber(): TransactionObject<string>

    numberValidatorsInSet(blockNumber: number | string): TransactionObject<string>

    setRegistry(registryAddress: string): TransactionObject<void>

    setSlashingIncentives(
      penalty: number | string,
      reward: number | string
    ): TransactionObject<void>

    getEpochSize(): TransactionObject<string>

    minQuorumSize(blockNumber: number | string): TransactionObject<string>

    fractionMulExp(
      aNumerator: number | string,
      aDenominator: number | string,
      bNumerator: number | string,
      bDenominator: number | string,
      exponent: number | string,
      _decimals: number | string
    ): TransactionObject<{
      0: string
      1: string
    }>

    transferOwnership(newOwner: string): TransactionObject<void>

    getParentSealBitmap(blockNumber: number | string): TransactionObject<string>

    initialize(
      registryAddress: string,
      _penalty: number | string,
      _reward: number | string
    ): TransactionObject<void>

    checkForDoubleSigning(
      signer: string,
      index: number | string,
      headerA: string | number[],
      headerB: string | number[]
    ): TransactionObject<string>

    slash(
      signer: string,
      index: number | string,
      headerA: string | number[],
      headerB: string | number[],
      groupMembershipHistoryIndex: number | string,
      validatorElectionLessers: string[],
      validatorElectionGreaters: string[],
      validatorElectionIndices: (number | string)[],
      groupElectionLessers: string[],
      groupElectionGreaters: string[],
      groupElectionIndices: (number | string)[]
    ): TransactionObject<void>
  }
  events: {
    SlashingIncentivesSet: ContractEvent<{
      penalty: string
      reward: string
      0: string
      1: string
    }>
    DoubleSigningSlashPerformed: ContractEvent<{
      validator: string
      blockNumber: string
      0: string
      1: string
    }>
    RegistrySet: ContractEvent<string>
    OwnershipTransferred: ContractEvent<{
      previousOwner: string
      newOwner: string
      0: string
      1: string
    }>
    allEvents: (options?: EventOptions, cb?: Callback<EventLog>) => EventEmitter
  }
}